refactor: merge simple functions
