--- packages/runtime/src/getInitialContext.ts
@@ -1,15 +1,0 @@
-import { createSearchParams } from 'react-router-dom';
-import type { InitialContext } from './types';
-
-export default function getInitialContext() {
-  const { href, origin, pathname, search } = window.location;
-  const path = href.replace(origin, '');
-  const query = Object.fromEntries(createSearchParams(search));
-  const initialContext: InitialContext = {
-    pathname,
-    path,
-    query,
-  };
-
-  return initialContext;
-}
